Re-export NDDSHOME variable in environment hook.
The `find_package(Connext)` function provided by connext_cmake_module
relies on the NDDSHOME variable being set.

In order for the connext_cmake_module package to be the only one which
needs to source the RTI environment script on the buildfarm the
environment hook must re-export it.

The implementation was inspired by https://github.com/ros2/rmw_opensplice/pull/233
but I noticed that the Batch file hook sources the upstream connext
script internally which would presumably set the NDDSHOME variable. So
the change to the Batch hook in this commit may not be needed.